# flask_app/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Products microservice",
    "description": "flask_app -- the products service",
    "version": "1.0.0"
  },
  "paths": {
    "/products": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "List every product",
        "responses": {
          "200": {
            "description": "All products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create a product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created, with an empty body"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Fetch one product by id",
        "responses": {
          "200": {
            "description": "The product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "No product with that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Update an existing product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated, with an empty body"
          },
          "404": {
            "description": "No product with that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "products"
        ],
        "summary": "Delete a product",
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "No product with that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Product 1"
          },
          "price": {
            "type": "number",
            "example": 10.99
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "required": [
          "name",
          "price"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Widget"
          },
          "price": {
            "type": "number",
            "example": 4.5
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "example": "Product not found"
          }
        }
      }
    }
  }
}